pytest/requirements.txt
pytest
pytest-assume
pyyaml
prettytable
onnx-simplifier